Add color 'theme' option
This adds the ability to choose a primer/css theme (i.e. light, dark,
etc.). To use this feature one can specify a 'primer.theme' in the
config file to set a theme site wide, or set 'theme' in the front matter
of a page to set the theme for that page.

Updating to the newer version of primer/css requires using Dart sass.
This adds 'sass_embedded' to the gemspec file as a dependency.